fixing execute_* option for NPB test script
The execute_* option will execute the specified benchmarks if this binaries
already exist in bin, and should compile and then execute them if the binaries
do not exist.

In contrast, the complete_* option will always recompile and execute all the
specified benchmarks.

Also, this modifies the clean and veryclean rules in the NPB makefile.